Resolve PR comments, rename chart, import in index, clean up
