+++ src/components/index.ts
@@ -10,6 +10,7 @@ import ESGTag from './Card/ESGTag'
 import NetAssetTag from './Card/NetAssetTag'
 // Charts
 import LandingDonutChart from './Charts/LandingDonutChart'
+import FinanceBrushChart from './Charts/FinanceBrushChart'
 // Displays
 import ListItem from './Displays/ListItem'
 import Carousel from './Displays/Carousel/Carousel'
@@ -42,6 +43,7 @@ export {
   Carousel,
   Select,
   LandingDonutChart,
+  FinanceBrushChart,
   CapitalizedTitle,
   RequestAccordion,
   SecondaryNavBar,
